requirements.txt
google-generativeai
openai
python-dotenv
requests
